Release 0.2.1: CSV BOM, headerRowIndex, ExcelDataType.TIME fix, and more
Breaking Changes:
- Cursor.getCurrentTotal() now returns long instead of int
- ExcelDataType.TIME now accepts LocalTime instead of LocalDateTime

New Features:
- Add UTF-8 BOM to CSV output for Excel Korean text compatibility
- Add ExcelReader.headerRowIndex() for files with metadata rows
- Add ExcelWriter.rowHeight() for configurable data row height
- Add CellData.resetDateFormats() / resetDateTimeFormats()

Bug Fixes:
- Fix CsvReadHandler exception double-wrapping
- Strip BOM from CSV header when reading

Improvements:
- Extract magic number to AUTO_WIDTH_SAMPLE_ROWS constant
- Fix method name typo: applyColumWidthAllSheets → applyColumnWidthAllSheets
- Update example code to match API changes
- Update README with new features and changelog

+++ example/src/main/java/io/github/dornol/excelkit/example/app/excel/BookExcelMapper.java
@@ -28,7 +28,7 @@ public class BookExcelMapper {
     public static ExcelHandler getHandler(Stream<BookDto> stream) {
         return new ExcelWriter<BookDto>(0xCC, 0xFF, 0x99)
                 .title("this is my title", 24, IndexedColors.DARK_BLUE)
-                .column("no", (rowData, cursor) -> cursor.getCurrentTotal()).type(ExcelDataType.INTEGER)
+                .column("no", (rowData, cursor) -> cursor.getCurrentTotal()).type(ExcelDataType.LONG)
                 .column("id", BookDto::id).type(ExcelDataType.LONG)
                 .column("title", BookDto::title)
                 .column("subtitle", BookDto::subtitle)
